When response.ok is False, executing response.json() will raise an opaque exception that doesn't give much information on what happened.
Setting response_json to None on those cases and passing the status_code
is better in terms of dealing with flux errors.

This is extreme useful when receiving multiple http 429 errors in a row.

Also, since exception message is sent to user when it happens, it's
better to preserve status code.